実装見直し
- Future は非同期状態で返却する
